v1.7.1 RC: hardening + logo fix
- mypy: 292→0 errors across 397 source files
- ruff: 354→39 cosmetic-only (0 bugs)
- New retry handler with exp backoff, jitter, timeout, error classification
- Removed dead code (cores/gateway/router.py, unused helpers)
- Fixed real bugs: Chinese literal, nonexistent enum attrs, coroutine await,
  bare except→Exception, missing raise...from, unused vars, E402 imports
- Tests: 331 passed, 2 xfailed (100%)
- Logo: cateye-logo SVGs copied to frontend/public/
- LoginPage.vue: replaced colored-letter placeholder with real SVG
- Activation.vue: added logo, removed unused Skeleton import

+++ cores/identity_vault.py
@@ -227,7 +227,7 @@ class IdentityVault:
 
         hours_since_check = 999
         if last_checked:
-            hours_since_check = (datetime.now(timezone.utc) - last_checked).total_seconds() / 3600
+            hours_since_check = int((datetime.now(timezone.utc) - last_checked).total_seconds() / 3600)
 
         state = data.get("session_state", "disconnected")
         has_creds = bool(data.get("encrypted_token") or data.get("encrypted_password"))
